remove redundant
fixes #5

God I'm dumb. I spent a fuckton of time trying to make sure they only run exactly when they need to. You know what? just go. It'll be fine. If performance takes a hit when I have a drink with a fuckton of ingredients, either we can revisit this, or maybe you should stop drinking from the kitchen sink